Finish up eslint setup
