+++ .npmignore
@@ -23,6 +23,7 @@ jest-puppeteer.config.js
 npm-debug.log*
 rollup.config.js
 tsconfig.json
-tslint.json
+.eslintrc.js
+prettier*
 yarn-debug.log*
 yarn-error.log*
