+++ test/compile/data/bin.test.ts
@@ -242,7 +242,7 @@ describe('compile/data/bin', () => {
     const binNodeA = new BinNode(parent, {foo: {bin: {}, field: 'foo', as: [['foo', 'foo_end']]}});
     const binNodeB = new BinNode(parent, {foo: {bin: {}, field: 'foo', as: [['bar', 'bar_end']]}});
 
-    binNodeA.merge(binNodeB, (s1, s2) => {});
+    binNodeA.merge(binNodeB, () => {});
     expect(binNodeA).toEqual(
       new BinNode(parent, {foo: {bin: {}, field: 'foo', as: [['foo', 'foo_end'], ['bar', 'bar_end']]}})
     );
@@ -253,7 +253,7 @@ describe('compile/data/bin', () => {
     const binNodeA = new BinNode(parent, {foo: {bin: {}, field: 'foo', as: [['foo', 'foo_end']]}});
     const binNodeB = new BinNode(parent, {foo: {bin: {}, field: 'foo', as: [['foo', 'foo_end'], ['bar', 'bar_end']]}});
 
-    binNodeA.merge(binNodeB, (s1, s2) => {});
+    binNodeA.merge(binNodeB, () => {});
     expect(binNodeA).toEqual(
       new BinNode(parent, {foo: {bin: {}, field: 'foo', as: [['foo', 'foo_end'], ['bar', 'bar_end']]}})
     );
@@ -275,7 +275,7 @@ describe('compile/data/bin', () => {
     const childA = new DataFlowNode(binNodeA);
     const childB = new DataFlowNode(binNodeB);
 
-    binNodeA.merge(binNodeB, (s1, s2) => {});
+    binNodeA.merge(binNodeB, () => {});
 
     expect(binNodeB.children.length).toEqual(0);
     expect(binNodeA.children.length).toEqual(2);
@@ -288,7 +288,7 @@ describe('compile/data/bin', () => {
     const binNodeA = new BinNode(parent, {foo: {bin: {}, field: 'foo', as: [['foo', 'foo_end']]}});
     const binNodeB = new BinNode(parent, {bar: {bin: {}, field: 'bar', as: [['bar', 'bar_end']]}});
 
-    binNodeA.merge(binNodeB, (s1, s2) => {});
+    binNodeA.merge(binNodeB, () => {});
     expect(binNodeA).toEqual(
       new BinNode(parent, {
         foo: {bin: {}, field: 'foo', as: [['foo', 'foo_end']]},
