+++ test/compile/data/optimize.test.ts
@@ -8,11 +8,8 @@ describe('compile/data/optimize', () => {
   describe('mergeParse', () => {
     it('should merge non-conflicting ParseNodes', () => {
       const root = new DataFlowNode(null, 'root');
-      // ts-ignore is used to suppress the noUnusedLocals error
-      // @ts-ignore
       const parse1 = new ParseNode(root, {a: 'number', b: 'string'});
-      // @ts-ignore
-      const parse2 = new ParseNode(root, {b: 'string', c: 'boolean'});
+      new ParseNode(root, {b: 'string', c: 'boolean'});
       const optimizer = new MergeParse();
       optimizer.run(parse1);
       expect(root.children.length).toEqual(1);
@@ -22,10 +19,8 @@ describe('compile/data/optimize', () => {
 
     it('should not merge conflicting ParseNodes', () => {
       const root = new DataFlowNode(null, 'root');
-      // @ts-ignore
       const parse1 = new ParseNode(root, {a: 'number', b: 'string'});
-      // @ts-ignore
-      const parse2 = new ParseNode(root, {a: 'boolean', d: 'date'});
+      new ParseNode(root, {a: 'boolean', d: 'date'});
       const optimizer = new MergeParse();
       optimizer.run(parse1);
       expect(root.children.length).toEqual(1);
@@ -41,10 +36,8 @@ describe('compile/data/optimize', () => {
       const source = new SourceNode(null);
       const parseOne = new ParseNode(source, {a: 'time', b: 'number'});
       const parseTwo = new ParseNode(source, {a: 'time', b: 'date'});
-      // @ts-ignore
-      const outputOne = new OutputNode(parseOne, 'foo', null, {foo: 1});
-      // @ts-ignore
-      const outputTwo = new OutputNode(parseTwo, 'bar', null, {bar: 1});
+      new OutputNode(parseOne, 'foo', null, {foo: 1});
+      new OutputNode(parseTwo, 'bar', null, {bar: 1});
 
       optimizeDataflow({sources: [source]} as any, null);
 
@@ -66,10 +59,8 @@ describe('compile/data/optimize', () => {
       const parseOne = new ParseNode(source, {a: 'time'});
       const parseTwo = new ParseNode(source, {b: 'number'});
       const parseThree = new ParseNode(parseTwo, {a: 'number'});
-      // @ts-ignore
-      const outputOne = new OutputNode(parseOne, 'foo', null, {foo: 1});
-      // @ts-ignore
-      const outputTwo = new OutputNode(parseThree, 'bar', null, {bar: 1});
+      new OutputNode(parseOne, 'foo', null, {foo: 1});
+      new OutputNode(parseThree, 'bar', null, {bar: 1});
 
       optimizeDataflow({sources: [source]} as any, null);
