+++ test/compile/data/parse.test.ts
@@ -93,7 +93,13 @@ describe('compile/data/parse', () => {
       const result = parseTransformArray(root, model, parse);
       expect(root.children[0] instanceof BinNode);
       expect(result instanceof TimeUnitNode);
-      expect(parse.combine()).toEqual({a: 'number', a_end: 'number', b: 'date', field: 'date'});
+      expect(parse.combine()).toEqual({
+        a: 'number',
+        // eslint-disable-next-line @typescript-eslint/camelcase
+        a_end: 'number',
+        b: 'date',
+        field: 'date'
+      });
     });
 
     it('should return a BinNode and a AggregateNode', () => {
