+++ test/compile/selection/interval.test.ts
@@ -474,15 +474,15 @@ describe('Interval Selections', () => {
       expect.arrayContaining([
         {
           name: 'one_modify',
-          update: `modify(\"one_store\", ${oneExpr})`
+          update: `modify("one_store", ${oneExpr})`
         },
         {
           name: 'two_modify',
-          update: `modify(\"two_store\", ${twoExpr})`
+          update: `modify("two_store", ${twoExpr})`
         },
         {
           name: 'thr_ee_modify',
-          update: `modify(\"thr_ee_store\", ${threeExpr})`
+          update: `modify("thr_ee_store", ${threeExpr})`
         }
       ])
     );
