+++ test/compile/selection/toggle.test.ts
@@ -86,11 +86,11 @@ describe('Toggle Selection Transform', () => {
       expect.arrayContaining([
         {
           name: 'one_modify',
-          update: `modify(\"one_store\", ${oneExpr})`
+          update: `modify("one_store", ${oneExpr})`
         },
         {
           name: 'two_modify',
-          update: `modify(\"two_store\", ${twoExpr})`
+          update: `modify("two_store", ${twoExpr})`
         }
       ])
     );
