+++ typings/vega-event-selector/index.d.ts
@@ -1,3 +1,3 @@
 declare module 'vega-event-selector' {
-  export function selector(selector: string, source: string): any[];
+  export function selector(selectorName: string, source: string): any[];
 }
